% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_simpop.R
\name{add_descendent_counts_to_tree}
\alias{add_descendent_counts_to_tree}
\title{Adds a vector "dcount" to a phylo object giving the number of descendent tips for each child node}
\usage{
add_descendent_counts_to_tree(tree)
}
\arguments{
\item{tree}{phylo}
}
\value{
phylo with dcount vector in parallel to edge.length
}
\description{
Adds a vector "dcount" to a phylo object giving the number of descendent tips for each child node
}
