% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{addCellCompartment}
\alias{addCellCompartment}
\title{Adds a new cell compartment to a simpop configuration}
\usage{
addCellCompartment(
  cfg,
  population,
  rate,
  ndriver = 0,
  basefit = 0,
  descr = NULL
)
}
\arguments{
\item{cfg}{list. simpop config}

\item{rate}{target population}

\item{ndriver}{integer. DEPRECATED}

\item{basefit}{numeric. DEPRECATED}

\item{descr}{character. Description}
}
\value{
list simpop config
}
\description{
Adds a new cell compartment to a simpop configuration
}
