% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{addDriverEvent}
\alias{addDriverEvent}
\title{Adds a driver event to a current simpop}
\usage{
addDriverEvent(tree, cfg, currentCompartment, fitness)
}
\arguments{
\item{tree}{\itemize{
\item simpop.
}}

\item{cfg}{\itemize{
\item List. simpop config.
}}

\item{currentCompartment}{\itemize{
\item integer. ID (value) of compartment to be added to.
}}
}
\description{
Adds a driver event to a current simpop
}
