% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{C_subsample_pop}
\alias{C_subsample_pop}
\title{Wrapper function for sub-sampling a simpop object}
\usage{
C_subsample_pop(tree, tips)
}
\arguments{
\item{tree}{simpop object.}

\item{tips}{character vector - names of tips to include in sub-sample}
}
\value{
simpop
}
\description{
Wrapper function for sub-sampling a simpop object
}
