% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped_sims.R
\name{continue_driver_process_sim}
\alias{continue_driver_process_sim}
\title{Restarts a multiple driver scenario}
\usage{
continue_driver_process_sim(insim, nyears, fitnessGen)
}
\arguments{
\item{insim}{\itemize{
\item Result of previous sim
}}

\item{nyears}{\itemize{
\item Total number of years to run the simulation
}}

\item{final_division_rate}{\itemize{
\item Rate of symmetric cell division once population equilibrium is reached.
}}

\item{target_pop_size}{\itemize{
\item Size of target population
}}

\item{drivers_per_year}{\itemize{
\item The expected number of drivers per year
}}

\item{bForceReseed}{\itemize{
\item Whether to reseed.
}}

\item{offset}{\itemize{
\item time offset for choosing seed.
}}
}
\value{
simpop object.
}
\description{
The user specifies the rate at which drivers are introduced and the distribution the selective coefficients are drawn from.
Note that unlike in \code{\link{run_selection_sim}} the drivers are allowed to stochastically die out.
The drivers at the specified rate with the gap between successive introductions exponentially distributed.
}
\examples{
fitnessGen=function(){
 trials=rexp(100,rate=30)
 idx=which(trials>0.05)
 if(length(idx)==0){
  fitnessGen()
 }else{
  trials[idx[1]]
 }
}
dps=run_driver_process_sim(0.1,1/(2*190),target_pop_size = 1e5,nyears = 30,fitness=fitnessGen,drivers_per_year = 5)
## Do stuff - subsample, plot trees etc and the restart to simulate the next 10 years:
dps2=continue_driver_process_sim(dps,40,fitnessGen = fitnessGen)
## Do stuff and simulation the next 10 years
dps2=continue_driver_process_sim(dps2,50,fitnessGen = fitnessGen)
}
