% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped_sims.R
\name{convert_annualS_to_s}
\alias{convert_annualS_to_s}
\title{Converts annual growth selection coefficient to per division selection coefficient}
\usage{
convert_annualS_to_s(S, divrate)
}
\arguments{
\item{S}{input annual selection coefficient}

\item{divrate}{division rate parameter specified in divisions per day.}
}
\value{
s : per division selection coefficient to be supplied to rsimpop
}
\description{
Converts annual growth selection coefficient to per division selection coefficient
}
