% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_annots.R
\name{get_all_node_children}
\alias{get_all_node_children}
\title{Gets all descendents of the specified node.}
\usage{
get_all_node_children(node, tree)
}
\arguments{
\item{node}{id of specified node}

\item{tree}{phylo. Tree}
}
\description{
TODO: Deprecate.. Does the same job as phytools::getDescendents
}
