% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_annots.R
\name{get_edge_info}
\alias{get_edge_info}
\title{Gets summary edge information specified by child node of an edge.}
\usage{
get_edge_info(pdx, tree, node)
}
\arguments{
\item{pdx}{list containing a data frame "details" with an entry "node" that maps to nodes in the tree.}

\item{tree}{phylo. Enhanced phylo - has coords data.frame specifying the coordinates of the edges.}

\item{node}{integer.}
}
\description{
Gets summary edge information specified by child node of an edge.
}
