% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{get_elapsed_time_tree}
\alias{get_elapsed_time_tree}
\title{Gets elapsed time tree with optional mutational aquisition modelling.
Uses the model burden ~ sum(pois(mutatrateperdiv))+pois(backgroundrate*duration)
If mutrateperdivision is not provide the function returns the ultrametric elapsed time tree.
Note: the tree that comes out of sim_pop has branch lengths=number of symmetric divisions (i.e. edge_length=ndivs)}
\usage{
get_elapsed_time_tree(
  tree,
  mutrateperdivision = NULL,
  backgroundrate = NULL,
  odf = 1
)
}
\arguments{
\item{tree}{simpop phylo. output from sim_pop or get_subsampled_tree.}

\item{mutrateperdivision}{numeric. Mutation rate per division}

\item{backgroundrate}{numeric. Negative Binomial or Poisson Distributed Mutation count}

\item{odf}{numeric. Overdispersion factor.}
}
\value{
simpop/phylo - with branch lengths defined as above (sometimes stochastic)
}
\description{
Gets elapsed time tree with optional mutational aquisition modelling.
Uses the model burden ~ sum(pois(mutatrateperdiv))+pois(backgroundrate*duration)
If mutrateperdivision is not provide the function returns the ultrametric elapsed time tree.
Note: the tree that comes out of sim_pop has branch lengths=number of symmetric divisions (i.e. edge_length=ndivs)
}
