% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_annots.R
\name{get_node_children}
\alias{get_node_children}
\title{Gets node id of immediate descendents of the specified nodes.}
\usage{
get_node_children(node, tree)
}
\arguments{
\item{node}{id of specified node}

\item{tree}{phylo.}
}
\description{
Will be empty for terminal nodes.
}
