% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_annots.R
\name{get_samples_in_clade}
\alias{get_samples_in_clade}
\title{Gets tip labels of tips that descend from the specified node}
\usage{
get_samples_in_clade(node, tree)
}
\arguments{
\item{node}{id of specified node}

\item{tree}{phylo.}
}
\description{
Gets tip labels of tips that descend from the specified node
}
