% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{get_subsampled_tree}
\alias{get_subsampled_tree}
\title{Gets subsampled tree - retaining events information appropriately}
\usage{
get_subsampled_tree(
  tree,
  N,
  tips = tree$edge[c(which(tree$state == 0 & tree$edge[, 2] <= length(tree$tip.label)),
    sample(which(tree$state != 0 & tree$edge[, 2] <= length(tree$tip.label)), N)), 2]
)
}
\arguments{
\item{tree}{simpop}

\item{N}{integer. Number of tips to subsample (ignored if tips is specified).}

\item{tips}{integer. node id of tips to subsample. Must include out-group.}
}
\value{
simpop object
}
\description{
Gets subsampled tree - retaining events information appropriately
}
\examples{
cfg=getDefaultConfig(target_pop_size =5e4,rate=0.1)
params=list(n_sim_days=2*365,b_stop_at_pop_size=1,b_stop_if_empty=0)
growthphase=sim_pop(NULL,params=params,cfg)
stree=get_subsampled_tree(growthphase,50)
}
