% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{get_subsampled_tree_fixed_compartmentcount}
\alias{get_subsampled_tree_fixed_compartmentcount}
\title{Gets subsampled tree - retaining events information appropriately}
\usage{
get_subsampled_tree_fixed_compartmentcount(tree, compartments, counts)
}
\arguments{
\item{tree}{simpop}

\item{compartments}{integer. compartment ids (see val in tree$cfg)}

\item{counts}{integer. number of cells to sample from corresponding compartment.}
}
\value{
simpop
}
\description{
Gets subsampled tree - retaining events information appropriately
}
