% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{get_tree_from_simpop}
\alias{get_tree_from_simpop}
\title{Converts simpop into an augmented APE phylo object of extant cells}
\usage{
get_tree_from_simpop(sim, b_check = FALSE)
}
\arguments{
\item{sim}{input simpop simulation}
}
\value{
ape tree
}
\description{
Converts simpop into an augmented APE phylo object of extant cells
}
