% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{getDefaultConfig}
\alias{getDefaultConfig}
\title{Gets a simple default configuration with one compartment}
\usage{
getDefaultConfig(
  target_pop_size,
  rate,
  descr = "cellType1",
  ndriver = 1,
  basefit = 0.1
)
}
\arguments{
\item{rate}{float - division rate in division per day}

\item{descr}{character - description of main compartment}

\item{ndriver}{integer - number of drivers - redundant now.}

\item{basefit}{float - base fitness redundant now.}

\item{targetPopSize}{integer - a negative number uses current time to 100th second resoultion.}
}
\value{
list config
}
\description{
Gets a simple default configuration with one compartment
}
