% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{initSimPop}
\alias{initSimPop}
\title{Initialises rsimpop}
\usage{
initSimPop(seed, bForce = FALSE)
}
\arguments{
\item{seed}{integer - a negative number uses current time to 100th second resoultion.}

\item{b}{boolean - whether to force re-initiliasation to specified seed.}
}
\description{
Initialises rsimpop
}
\examples{
initSimPop(-1)
}
