% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_annots.R
\name{node_labels}
\alias{node_labels}
\title{Adds node labels to an existing tree plot}
\usage{
node_labels(tree, col = "blue", cex = 1, b_include_tips = FALSE)
}
\arguments{
\item{tree}{enhanced phylo returned by plot_tree}

\item{col}{character. Colour specification for labels}

\item{cex}{numeric.  Size of labels}

\item{b_include_tips}{boolean. Wether to include tips
#' @return simpop object}
}
\description{
Adds node labels to an existing tree plot
}
\examples{
test=rtree(20)
tree=plot_tree(test,cex.terminal=0.5,bars=sapply(sprintf("s\%d",seq(2,20,2)),function(x) runif(1)))
node_labels(tree)

}
