% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{plot.simpop}
\alias{plot.simpop}
\title{Plot simpop population trajectory}
\usage{
\method{plot}{simpop}(simpop, ...)
}
\arguments{
\item{simpop}{integer - a negative number uses current time to 100th second resoultion.}
}
\description{
Plot simpop population trajectory
}
