% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{plot_compartment_trajectory}
\alias{plot_compartment_trajectory}
\title{Plot simpop compartment level population trajectory}
\usage{
plot_compartment_trajectory(simpop)
}
\arguments{
\item{simpop}{integer - a negative number uses current time to 100th second resoultion.}
}
\value{
ggplot
}
\description{
Plot simpop compartment level population trajectory
}
