% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_annots.R
\name{plot_tree}
\alias{plot_tree}
\title{Plots an ape phylo tree}
\usage{
plot_tree(
  tree,
  direction = "down",
  cex.label = 1,
  offset = 0,
  b_do_not_plot = FALSE,
  lwd = 1,
  bars = NULL,
  default_edge_color = "darkgrey",
  ymax = NULL,
  cex.terminal.dots = 0,
  mar = NULL,
  b.add.scale = TRUE,
  left.margin.prop = 0.1
)
}
\arguments{
\item{tree}{phylo/simpop}

\item{direction}{character. Only "down" is currently supported.}

\item{cex.label}{numeric. Size of tip labels (0 results in no tip labels)}

\item{offset}{numeric (0-1). Fractional offset of tip labels from branch terminus}

\item{b_do_not_plot}{Boolean.}

\item{lwd}{numeric. Line width}

\item{bars}{numeric. vector with named elements that should all match tip labels - adds barplot under tree.}

\item{default_edge_color}{colour.}

\item{ymax}{numeric max y}

\item{cex.terminal.dots}{numeric - size of terminal dots}

\item{mar}{numeric vector Override default resetting of mar as in par("mar")}

\item{b.add.scale}{Boolean - whether to add scale axis}

\item{left.margin.prop}{numeric - leaves this fraction empty on left}
}
\value{
simpop object
}
\description{
This is similar to APE plot.phylo but returns useful information on each branch to facilitate annotation of tree.
Be careful not to unintentionally supply population scale simpop
}
\examples{
test=rtree(20)
tree=plot_tree(test,cex.terminal=0.5,bars=sapply(sprintf("s\%d",seq(2,20,2)),function(x) runif(1)))
node_labels(tree)
}
