% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_simpop.R
\name{plot_tree_events}
\alias{plot_tree_events}
\title{Plots simpop events on a the tree}
\usage{
plot_tree_events(tree, legpos = "topleft", ...)
}
\arguments{
\item{tree}{phylo/simpop}

\item{legpos}{legend position}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> . Parameters that are passed to plot_tree}
}
\value{
simpop object
}
\description{
Plots simpop events on a the tree
}
\examples{
cfg=getDefaultConfig(target_pop_size =5e4,rate=0.1)
params=list(n_sim_days=2*365,b_stop_at_pop_size=1,b_stop_if_empty=0)
growthphase=sim_pop(NULL,params=params,cfg)
stree=get_subsampled_tree(growthphase,50)
plot_tree_events(stree)
}
