% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped_sims.R
\name{run_driver_process_sim}
\alias{run_driver_process_sim}
\title{Runs a multiple driver scenario}
\usage{
run_driver_process_sim(
  initial_division_rate = 0.1,
  final_division_rate = 1/365,
  target_pop_size = 1e+05,
  drivers_per_year = 0.1,
  nyears = 40,
  fitnessGen = function() {     0 },
  bForceReseed = FALSE,
  offset = 0
)
}
\arguments{
\item{initial_division_rate}{\itemize{
\item Rate of symmetric cell division during development
}}

\item{final_division_rate}{\itemize{
\item Rate of symmetric cell division once population equilibrium is reached.
}}

\item{target_pop_size}{\itemize{
\item Size of target population
}}

\item{drivers_per_year}{\itemize{
\item The expected number of drivers per year
}}

\item{nyears}{\itemize{
\item Total number of years to run the simulation
}}

\item{fitnessGen}{\itemize{
\item Function that returns a single value for the selective coefficient
}}

\item{bForceReseed}{\itemize{
\item Whether to reseed.
}}

\item{offset}{\itemize{
\item time offset for choosing seed.
}}
}
\value{
simpop object.
}
\description{
The user specifies the rate at which drivers are introduced and the distribution of the selective coefficients are drawn from.
Note that unlike in \code{\link{run_selection_sim}} the drivers are allowed to stochastically die out.
The drivers at the specified rate with the gap between successive introductions exponentially distributed.
}
\examples{
fitnessGen=function(){
 trials=rexp(100,rate=30)
 idx=which(trials>0.05)
 if(length(idx)==0){
  fitnessGen()
 }else{
  trials[idx[1]]
 }
}
dps=run_driver_process_sim(0.1,1/(2*190),target_pop_size = 1e5,nyears = 30,fitness=fitnessGen,drivers_per_year = 5)

}
