% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped_sims.R
\name{run_neutral_sim}
\alias{run_neutral_sim}
\title{Runs a simple neutral simulation}
\usage{
run_neutral_sim(
  initial_division_rate,
  final_division_rate,
  target_pop_size = 1e+05,
  nyears = 40
)
}
\arguments{
\item{initial_division_rate}{\itemize{
\item Rate of symmetric cell division during development
}}

\item{final_division_rate}{\itemize{
\item Rate of symmetric cell division once population equilibrium is reached.
}}

\item{target_pop_size}{\itemize{
\item Size of target population
}}

\item{nyears}{\itemize{
\item Total number of years to run the simulation
}}
}
\value{
simpop object.
}
\description{
Runs a simple neutral simulation
}
\examples{
neutsim=run_neutral_sim(0.05,1/365,target_pop_size = 5e4,nyears=80)
}
