% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped_sims.R
\name{run_selection_sim}
\alias{run_selection_sim}
\title{Runs a single compartment/single driver scenario}
\usage{
run_selection_sim(
  initial_division_rate = 0.1,
  final_division_rate = 1/365,
  target_pop_size = 1e+05,
  nyears_driver_acquisition = 15,
  nyears = 40,
  fitness = 0.2,
  minprop = 0.001,
  mindriver = 1,
  maxtry = 40
)
}
\arguments{
\item{initial_division_rate}{\itemize{
\item Rate of symmetric cell division during development
}}

\item{final_division_rate}{\itemize{
\item Rate of symmetric cell division once population equilibrium is reached.
}}

\item{target_pop_size}{\itemize{
\item Size of target population
}}

\item{nyears_driver_acquisition}{\itemize{
\item When driver is acquired.
}}

\item{nyears}{\itemize{
\item Total number of years to run the simulation
}}

\item{fitness}{\itemize{
\item relative fitness advantage.  Cells carrying this divide at a rate=(1+fitness)*baserate
}}

\item{minprop}{\itemize{
\item Minimum aberrant cell fraction at nyears for driver to be regarded as have "taken"
}}

\item{maxtry}{\itemize{
\item Maximum number of attempts to introduce a driver
}}
}
\value{
simpop object.
}
\description{
run_selection_sim models the evolution of a population of cells with a selective advantage given to a randomly
chosen cell, once the population size reaches the specified equilibrium size.
}
\details{
The selective advantage is introduced as a driver event. The driver is subject to stochastic
extinction and so is repeatedly dropped in until it "takes".The prevailing state when the driver is introduced is
saved and reinstated with each attempted introduction.
}
\examples{
selsim=run_selection_sim(0.05,1/365,target_pop_size = 5e4,nyears = 50,fitness=0.3)
}
