% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped_sims.R
\name{run_transient_selection}
\alias{run_transient_selection}
\title{Runs a single compartment/single driver scenario with transient selection}
\usage{
run_transient_selection(
  initial_division_rate,
  final_division_rate,
  target_pop_size = 1e+05,
  nyears_driver_acquisition = 15,
  nyears_transient_end = 30,
  nyears = 40,
  fitness = 0.2,
  minprop = 0.05
)
}
\arguments{
\item{initial_division_rate}{\itemize{
\item Rate of symmetric cell division during development
}}

\item{final_division_rate}{\itemize{
\item Rate of symmetric cell division once population equilibrium is reached.
}}

\item{target_pop_size}{\itemize{
\item Size of target population
}}

\item{nyears_driver_acquisition}{\itemize{
\item When driver is acquired.
}}

\item{nyears_transient_end}{\itemize{
\item When selective advantage is set to 0.
}}

\item{nyears}{\itemize{
\item Total number of years to run the simulation
}}

\item{fitness}{\itemize{
\item relative fitness advantage.  Cells carrying this divide at a rate=(1+fitness)*baserate
}}

\item{minprop}{\itemize{
\item Minimum aberrant cell fraction at nyears for driver to be regarded as have "taken"
}}

\item{maxtry}{\itemize{
\item Maximum number of attempts to introduce a driver
}}
}
\value{
simpop object.
}
\description{
run_transient_selection models the evolution of a population of cells with a transient selective advantage given to a randomly
chosen cell, once the population size reaches the specified equilibrium size.
}
\details{
This function is similar to the run_selection_sim() function. However, the driver is switch off after the time specified in the nyears_transient_end argument.
}
\examples{
tselsim=run_transient_selection(0.05,1/365,target_pop_size = 5e4,nyears_driver_acquisition=15,
nyears_transient_end=30,nyears=50,fitness=0.5)
}
