% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{combine_simpops}
\alias{combine_simpops}
\title{Concatenates simpops (maintains population size trajectory and drivercount trajectory)
Concatenates timestamps and population size}
\usage{
combine_simpops(simpop1, simpop2)
}
\arguments{
\item{simpop1}{simpop from initial run}

\item{simpop2}{simpop from next run}
}
\value{
simpop/phylo - with branch lengths defined as above (sometimes stochastic)
}
\description{
Concatenates simpops (maintains population size trajectory and drivercount trajectory)
Concatenates timestamps and population size
}
